+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG.md
 
+## 0.0.7 - 2023-09-10
+- Rename outerBorderColor to borderColor
+- Rename outerBorderWidth to borderWidth
+
 ## 0.0.6 - 2023-09-09
 - Add optionally displaying a border